View will not capture clickEvent by default.
So that, the event will pass to parent.

When we click on child view, it will capture the event, while
parent could not receive it.
However, sometime, the parent wishes to capture the event, and child
don't care about it. This event should be sent to parent, not child.

To fix it, let onclickInternal by default false. So that the event will
be sent to parent, if child don't explicitly deal it.