Cache: Add Layout Cache on SimpleView
This is a simple implementation.

Will try to avoid re-calculation in measure, layout.
If you want to force-update measure, layout of a view,
you shall set this view isDirty, and it will do so in
next drawToCanvas().

Still we have more things to do, like:
what if we force-update a view, should we update the
parent ? Should we update the other sibling views ?